@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,7 +29,7 @@ dependencies:
   firebase_auth: ^0.20.0+1
   cloud_firestore: ^0.16.0
   flutter_easyloading: ^2.2.2
-  geolocator: ^6.1.13
+  geolocator: ^7.0.0-nullsafety.5
   geocoding: ^1.0.5
   permission_handler: ^5.0.1+1
   emojis: ^0.9.3
